requirements.txt
numpy
open3d
scipy
torch
torchvision
trimesh
warp-lang
